feature: Extracts engine domain contracts into a types module
